requirements.txt
# Python dependencies
python-binance
python-dotenv
pandas
